Moved files into src directory
